Prepare a first pre-release
